refactor(vsock-guest): rename shell command helpers
