+++ crates/vsock-guest/src/lib.rs
@@ -10,12 +10,12 @@ mod command;
 mod connection;
 mod drain;
 mod error;
-mod exec;
 mod handlers;
 mod log;
 mod monitor;
 mod process;
 mod quiesce;
+mod shell_command;
 mod shutdown;
 mod threading;
 mod user;
