+++ crates/vsock-guest/src/process.rs
@@ -1,4 +1,5 @@
-use std::process::{Child, ExitStatus};
+use std::io;
+use std::process::{Child, Command, ExitStatus};
 
 use crate::log::log;
 
@@ -14,6 +15,22 @@ pub(crate) fn extract_exit_code(status: ExitStatus) -> i32 {
 #[cfg(not(unix))]
 pub(crate) fn extract_exit_code(status: ExitStatus) -> i32 {
     status.code().unwrap_or(1)
+}
+
+/// Spawn a command as the leader of a new process group on Unix.
+///
+/// Timeout killing targets the process group by child PID, so every child path
+/// that uses the shared wait helpers must preserve this spawn invariant.
+pub(crate) fn spawn_in_own_process_group(command: &mut Command) -> io::Result<Child> {
+    #[cfg(unix)]
+    {
+        use std::os::unix::process::CommandExt;
+        command.process_group(0).spawn()
+    }
+    #[cfg(not(unix))]
+    {
+        command.spawn()
+    }
 }
 
 /// Parse ppid and pgid from a `/proc/[pid]/stat` line.
@@ -155,7 +172,46 @@ impl Drop for ChildReapGuard {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::process::Command;
+    use std::os::unix::fs::PermissionsExt;
+    use std::path::PathBuf;
+    use std::process::{Command, Stdio};
+    use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
+
+    use crate::wait::{WaitOutcome, wait_with_kill_timeout};
+
+    struct TempDirGuard(PathBuf);
+
+    impl Drop for TempDirGuard {
+        fn drop(&mut self) {
+            let _ = std::fs::remove_dir_all(&self.0);
+        }
+    }
+
+    fn temp_dir(label: &str) -> (PathBuf, TempDirGuard) {
+        let dir = std::env::temp_dir().join(format!(
+            "vsock-guest-{label}-{}-{}",
+            std::process::id(),
+            SystemTime::now()
+                .duration_since(UNIX_EPOCH)
+                .unwrap()
+                .as_nanos()
+        ));
+        std::fs::create_dir_all(&dir).unwrap();
+        std::fs::set_permissions(&dir, std::fs::Permissions::from_mode(0o700)).unwrap();
+        let guard = TempDirGuard(dir.clone());
+        (dir, guard)
+    }
+
+    fn wait_for_path(path: &std::path::Path, timeout: Duration) -> bool {
+        let deadline = Instant::now() + timeout;
+        while Instant::now() < deadline {
+            if path.exists() {
+                return true;
+            }
+            std::thread::sleep(Duration::from_millis(10));
+        }
+        path.exists()
+    }
 
     #[test]
     fn extract_exit_code_success() {
@@ -251,14 +307,112 @@ mod tests {
 
     #[cfg(target_os = "linux")]
     fn wait_until_process_is_gone_or_zombie(pid: i32) -> bool {
-        let deadline = std::time::Instant::now() + std::time::Duration::from_secs(1);
-        while std::time::Instant::now() < deadline {
+        let deadline = Instant::now() + Duration::from_secs(1);
+        while Instant::now() < deadline {
             if process_is_gone_or_zombie(pid) {
                 return true;
             }
-            std::thread::sleep(std::time::Duration::from_millis(10));
+            std::thread::sleep(Duration::from_millis(10));
         }
         process_is_gone_or_zombie(pid)
+    }
+
+    #[cfg(target_os = "linux")]
+    fn open_pidfd(pid: libc::pid_t) -> std::io::Result<std::os::fd::OwnedFd> {
+        use std::os::fd::FromRawFd;
+
+        // SAFETY: `pidfd_open` does not dereference user pointers. On success
+        // it returns a new file descriptor owned by this process.
+        let fd = unsafe { libc::syscall(libc::SYS_pidfd_open, pid, 0) };
+        if fd < 0 {
+            return Err(std::io::Error::last_os_error());
+        }
+
+        // SAFETY: `fd` is a fresh descriptor returned by `pidfd_open` above.
+        Ok(unsafe { std::os::fd::OwnedFd::from_raw_fd(fd as std::os::fd::RawFd) })
+    }
+
+    #[cfg(target_os = "linux")]
+    fn wait_for_pidfd_exit(
+        pidfd: &std::os::fd::OwnedFd,
+        timeout: Duration,
+    ) -> std::io::Result<bool> {
+        let deadline = Instant::now() + timeout;
+        loop {
+            let remaining = deadline.saturating_duration_since(Instant::now());
+            if remaining.is_zero() {
+                return Ok(false);
+            }
+            let timeout_ms = remaining.as_millis().clamp(1, i32::MAX as u128) as i32;
+            let mut pollfd = libc::pollfd {
+                fd: std::os::fd::AsRawFd::as_raw_fd(pidfd),
+                events: libc::POLLIN,
+                revents: 0,
+            };
+            // SAFETY: `pollfd` points to one initialized descriptor entry.
+            let result = unsafe { libc::poll(&mut pollfd, 1, timeout_ms) };
+            if result > 0 {
+                let revents = pollfd.revents;
+                if revents & (libc::POLLERR | libc::POLLNVAL) != 0 {
+                    return Err(std::io::Error::other("pidfd became invalid while polling"));
+                }
+                if revents & (libc::POLLIN | libc::POLLHUP) != 0 {
+                    return Ok(true);
+                }
+                return Err(std::io::Error::other(format!(
+                    "unexpected pidfd poll revents: {revents:#x}"
+                )));
+            }
+            if result == 0 {
+                return Ok(false);
+            }
+            let err = std::io::Error::last_os_error();
+            if err.kind() != std::io::ErrorKind::Interrupted {
+                return Err(err);
+            }
+        }
+    }
+
+    #[cfg(target_os = "linux")]
+    fn kill_spawned_child(child: &mut Option<Child>) {
+        if let Some(child) = child.take() {
+            kill_and_reap_child(child);
+        }
+    }
+
+    #[cfg(target_os = "linux")]
+    fn signal_pidfd(pidfd: &std::os::fd::OwnedFd, signal: libc::c_int) -> std::io::Result<()> {
+        // SAFETY: best-effort cleanup of a test-owned background process.
+        let result = unsafe {
+            libc::syscall(
+                libc::SYS_pidfd_send_signal,
+                std::os::fd::AsRawFd::as_raw_fd(pidfd),
+                signal,
+                std::ptr::null::<libc::siginfo_t>(),
+                0,
+            )
+        };
+        if result == 0 {
+            return Ok(());
+        }
+        let err = std::io::Error::last_os_error();
+        if err.raw_os_error() == Some(libc::ESRCH) {
+            return Ok(());
+        }
+        Err(err)
+    }
+
+    #[cfg(target_os = "linux")]
+    fn kill_pidfd_and_wait(pidfd: &std::os::fd::OwnedFd) -> std::io::Result<()> {
+        signal_pidfd(pidfd, libc::SIGKILL)?;
+        if wait_for_pidfd_exit(pidfd, Duration::from_secs(1))? {
+            return Ok(());
+        }
+
+        Err(std::io::Error::new(
+            std::io::ErrorKind::TimedOut,
+            "timed out waiting for pidfd process to exit after SIGKILL",
+        ))
     }
 
     #[cfg(target_os = "linux")]
@@ -266,7 +420,6 @@ mod tests {
     fn kill_and_reap_child_kills_group_after_direct_child_exits() {
         use std::io::{BufRead, BufReader};
         use std::os::unix::process::CommandExt;
-        use std::process::Stdio;
 
         let mut command = Command::new("sh");
         command
@@ -288,6 +441,77 @@ mod tests {
         if !wait_until_process_is_gone_or_zombie(background_pid) {
             let _ = unsafe { libc::kill(background_pid, libc::SIGKILL) };
             panic!("background pid {background_pid} should have been killed");
+        }
+    }
+
+    #[cfg(target_os = "linux")]
+    #[test]
+    fn spawn_in_own_process_group_timeout_kills_background_child() {
+        let (dir, _guard) = temp_dir("pg");
+        let ready = dir.join("ready");
+        let background_pid = dir.join("background-pid");
+
+        let mut command = Command::new("sh");
+        command
+            .arg("-c")
+            .arg(
+                "trap '' HUP; tail -f /dev/null & echo $! > \"$PIDFILE\"; \
+                 touch \"$READY\"; wait",
+            )
+            .env("READY", &ready)
+            .env("PIDFILE", &background_pid)
+            .stdout(Stdio::null())
+            .stderr(Stdio::null());
+
+        let mut child = Some(spawn_in_own_process_group(&mut command).unwrap());
+        if !wait_for_path(&ready, Duration::from_secs(2)) {
+            kill_spawned_child(&mut child);
+            panic!("background child should be started before timeout kill is tested");
+        }
+        let background_pid_text = match std::fs::read_to_string(&background_pid) {
+            Ok(pid) => pid,
+            Err(e) => {
+                kill_spawned_child(&mut child);
+                panic!("failed to read background pid: {e}");
+            }
+        };
+        let background_pid: libc::pid_t = match background_pid_text.trim().parse() {
+            Ok(pid) => pid,
+            Err(e) => {
+                kill_spawned_child(&mut child);
+                panic!("failed to parse background pid {background_pid_text:?}: {e}");
+            }
+        };
+        let background_pidfd = match open_pidfd(background_pid) {
+            Ok(pidfd) => pidfd,
+            Err(e) => {
+                kill_spawned_child(&mut child);
+                panic!("failed to open pidfd for pid {background_pid}: {e}");
+            }
+        };
+
+        let outcome = wait_with_kill_timeout(child.take().unwrap(), 100);
+        if !matches!(outcome, WaitOutcome::TimedOut) {
+            kill_pidfd_and_wait(&background_pidfd)
+                .unwrap_or_else(|e| panic!("failed to clean up background pidfd: {e}"));
+            panic!("expected timeout kill to return WaitOutcome::TimedOut");
+        }
+
+        match wait_for_pidfd_exit(&background_pidfd, Duration::from_secs(2)) {
+            Ok(true) => {}
+            Ok(false) => {
+                kill_pidfd_and_wait(&background_pidfd)
+                    .unwrap_or_else(|e| panic!("failed to clean up background pidfd: {e}"));
+                panic!(
+                    "timeout kill should terminate background pid {background_pid} in the process group"
+                );
+            }
+            Err(e) => {
+                let cleanup = kill_pidfd_and_wait(&background_pidfd);
+                panic!(
+                    "failed to wait for background pid {background_pid} exit: {e}; cleanup={cleanup:?}"
+                );
+            }
         }
     }
 }
