+++ crates/vsock-guest/src/shell_command.rs
@@ -14,7 +14,7 @@ const ENV_SCRIPT_SUFFIX: &str = ".sh";
 const ENV_SCRIPT_STALE_AFTER: Duration = Duration::from_secs(60 * 60);
 const CHOWN_UNCHANGED_UID: libc::uid_t = !0;
 
-fn get_exec_user() -> Option<&'static str> {
+fn shell_command_user() -> Option<&'static str> {
     #[cfg(any(debug_assertions, feature = "test-support"))]
     {
         None
@@ -40,8 +40,8 @@ fn shell_escape_value(val: &str) -> String {
 /// In production-style builds, non-sudo commands run through `su - user`.
 /// In debug/test-support builds, local tests run as the current user unless
 /// `sudo` explicitly requests elevation through `sudo sh -c`.
-pub(crate) fn build_exec_command(command: &str, sudo: bool) -> Command {
-    match get_exec_user() {
+pub(crate) fn build_shell_command(command: &str, sudo: bool) -> Command {
+    match shell_command_user() {
         Some(user) => {
             if sudo {
                 // Release: already root — run directly
@@ -102,12 +102,12 @@ impl Drop for EnvScriptGuard {
     }
 }
 
-pub(crate) struct PreparedExecCommand {
+pub(crate) struct PreparedShellCommand {
     pub(crate) command: Command,
     pub(crate) env_script: Option<EnvScriptGuard>,
 }
 
-pub(crate) struct SpawnedCommand {
+pub(crate) struct SpawnedShellCommand {
     pub(crate) child: Child,
     pub(crate) env_script: Option<EnvScriptGuard>,
 }
@@ -128,7 +128,7 @@ fn default_env_script_dir() -> PathBuf {
 }
 
 fn format_env_key_for_log(key: &str) -> String {
-    truncate_preview(&key.escape_debug().to_string())
+    truncate_command_preview(&key.escape_debug().to_string())
 }
 
 fn compare_env_diagnostic_entries(
@@ -434,7 +434,7 @@ fn create_env_script_in_dir(
 
         let result = (|| -> io::Result<()> {
             file.write_all(script.as_bytes())?;
-            if effective_uid() == 0 && !sudo && get_exec_user().is_some() {
+            if effective_uid() == 0 && !sudo && shell_command_user().is_some() {
                 // Keep the per-run directory and script root-owned. The
                 // sandbox user only gets group read/traverse access; if it
                 // owned either path, an existing same-UID process could
@@ -484,14 +484,14 @@ fn script_invocation(path: &Path) -> io::Result<String> {
     Ok(format!("/bin/bash {}", shell_escape_value(path)))
 }
 
-pub(crate) fn build_exec_command_with_env(
+pub(crate) fn build_shell_command_with_env(
     command: &str,
     env: &[(&str, &str)],
     sudo: bool,
-) -> io::Result<PreparedExecCommand> {
+) -> io::Result<PreparedShellCommand> {
     if env.is_empty() {
-        return Ok(PreparedExecCommand {
-            command: build_exec_command(command, sudo),
+        return Ok(PreparedShellCommand {
+            command: build_shell_command(command, sudo),
             env_script: None,
         });
     }
@@ -501,14 +501,14 @@ pub(crate) fn build_exec_command_with_env(
         .path()
         .ok_or_else(|| io::Error::other("env script path missing"))?;
     let invocation = script_invocation(script_path)?;
-    Ok(PreparedExecCommand {
-        command: build_exec_command(&invocation, sudo),
+    Ok(PreparedShellCommand {
+        command: build_shell_command(&invocation, sudo),
         env_script: Some(env_script),
     })
 }
 
 /// Truncate a command string for logging, preserving UTF-8 boundaries
-pub(crate) fn truncate_preview(s: &str) -> String {
+pub(crate) fn truncate_command_preview(s: &str) -> String {
     if s.len() <= COMMAND_PREVIEW_MAX_LEN {
         return s.to_string();
     }
@@ -520,36 +520,20 @@ pub(crate) fn truncate_preview(s: &str) -> String {
     format!("{}...", &s[..end])
 }
 
-/// Spawn a command as the leader of a new process group on Unix.
-///
-/// Timeout killing targets the process group by child PID, so every child path
-/// that uses the shared wait helpers must preserve this spawn invariant.
-pub(crate) fn spawn_in_own_process_group(command: &mut Command) -> io::Result<Child> {
-    #[cfg(unix)]
-    {
-        use std::os::unix::process::CommandExt;
-        command.process_group(0).spawn()
-    }
-    #[cfg(not(unix))]
-    {
-        command.spawn()
-    }
-}
-
-/// Spawn `command` with stdout/stderr piped — used by both buffered exec and
-/// streaming spawned processes.
-pub(crate) fn spawn_with_pipes(
+/// Spawn a shell command with stdout/stderr piped for buffered and streaming
+/// spawned processes.
+pub(crate) fn spawn_shell_command_with_pipes(
     command: &str,
     env: &[(&str, &str)],
     sudo: bool,
-) -> io::Result<SpawnedCommand> {
-    let PreparedExecCommand {
+) -> io::Result<SpawnedShellCommand> {
+    let PreparedShellCommand {
         mut command,
         env_script,
-    } = build_exec_command_with_env(command, env, sudo)?;
+    } = build_shell_command_with_env(command, env, sudo)?;
     command.stdout(Stdio::piped()).stderr(Stdio::piped());
-    let child = spawn_in_own_process_group(&mut command)?;
-    Ok(SpawnedCommand { child, env_script })
+    let child = crate::process::spawn_in_own_process_group(&mut command)?;
+    Ok(SpawnedShellCommand { child, env_script })
 }
 
 #[cfg(test)]
@@ -557,82 +541,13 @@ mod tests {
     use super::*;
     use std::path::PathBuf;
     use std::sync::{Arc, Barrier};
-    use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
-
-    use crate::wait::{WaitOutcome, wait_with_kill_timeout};
+    use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
     struct TempDirGuard(PathBuf);
 
     impl Drop for TempDirGuard {
         fn drop(&mut self) {
             let _ = std::fs::remove_dir_all(&self.0);
-        }
-    }
-
-    fn wait_for_path(path: &std::path::Path, timeout: Duration) -> bool {
-        let deadline = Instant::now() + timeout;
-        while Instant::now() < deadline {
-            if path.exists() {
-                return true;
-            }
-            std::thread::sleep(Duration::from_millis(10));
-        }
-        path.exists()
-    }
-
-    #[cfg(target_os = "linux")]
-    fn open_pidfd(pid: libc::pid_t) -> std::io::Result<std::os::fd::OwnedFd> {
-        use std::os::fd::FromRawFd;
-
-        // SAFETY: `pidfd_open` does not dereference user pointers. On success
-        // it returns a new file descriptor owned by this process.
-        let fd = unsafe { libc::syscall(libc::SYS_pidfd_open, pid, 0) };
-        if fd < 0 {
-            return Err(std::io::Error::last_os_error());
-        }
-
-        // SAFETY: `fd` is a fresh descriptor returned by `pidfd_open` above.
-        Ok(unsafe { std::os::fd::OwnedFd::from_raw_fd(fd as std::os::fd::RawFd) })
-    }
-
-    #[cfg(target_os = "linux")]
-    fn wait_for_pidfd_exit(
-        pidfd: &std::os::fd::OwnedFd,
-        timeout: Duration,
-    ) -> std::io::Result<bool> {
-        let deadline = Instant::now() + timeout;
-        loop {
-            let remaining = deadline.saturating_duration_since(Instant::now());
-            if remaining.is_zero() {
-                return Ok(false);
-            }
-            let timeout_ms = remaining.as_millis().clamp(1, i32::MAX as u128) as i32;
-            let mut pollfd = libc::pollfd {
-                fd: std::os::fd::AsRawFd::as_raw_fd(pidfd),
-                events: libc::POLLIN,
-                revents: 0,
-            };
-            // SAFETY: `pollfd` points to one initialized descriptor entry.
-            let result = unsafe { libc::poll(&mut pollfd, 1, timeout_ms) };
-            if result > 0 {
-                let revents = pollfd.revents;
-                if revents & (libc::POLLERR | libc::POLLNVAL) != 0 {
-                    return Err(std::io::Error::other("pidfd became invalid while polling"));
-                }
-                if revents & (libc::POLLIN | libc::POLLHUP) != 0 {
-                    return Ok(true);
-                }
-                return Err(std::io::Error::other(format!(
-                    "unexpected pidfd poll revents: {revents:#x}"
-                )));
-            }
-            if result == 0 {
-                return Ok(false);
-            }
-            let err = std::io::Error::last_os_error();
-            if err.kind() != std::io::ErrorKind::Interrupted {
-                return Err(err);
-            }
         }
     }
 
@@ -649,48 +564,6 @@ mod tests {
         std::fs::set_permissions(&dir, std::fs::Permissions::from_mode(0o700)).unwrap();
         let guard = TempDirGuard(dir.clone());
         (dir, guard)
-    }
-
-    #[cfg(target_os = "linux")]
-    fn kill_spawned_child(child: &mut Option<Child>) {
-        if let Some(child) = child.take() {
-            crate::process::kill_and_reap_child(child);
-        }
-    }
-
-    #[cfg(target_os = "linux")]
-    fn signal_pidfd(pidfd: &std::os::fd::OwnedFd, signal: libc::c_int) -> std::io::Result<()> {
-        // SAFETY: best-effort cleanup of a test-owned background process.
-        let result = unsafe {
-            libc::syscall(
-                libc::SYS_pidfd_send_signal,
-                std::os::fd::AsRawFd::as_raw_fd(pidfd),
-                signal,
-                std::ptr::null::<libc::siginfo_t>(),
-                0,
-            )
-        };
-        if result == 0 {
-            return Ok(());
-        }
-        let err = std::io::Error::last_os_error();
-        if err.raw_os_error() == Some(libc::ESRCH) {
-            return Ok(());
-        }
-        Err(err)
-    }
-
-    #[cfg(target_os = "linux")]
-    fn kill_pidfd_and_wait(pidfd: &std::os::fd::OwnedFd) -> std::io::Result<()> {
-        signal_pidfd(pidfd, libc::SIGKILL)?;
-        if wait_for_pidfd_exit(pidfd, Duration::from_secs(1))? {
-            return Ok(());
-        }
-
-        Err(std::io::Error::new(
-            std::io::ErrorKind::TimedOut,
-            "timed out waiting for pidfd process to exit after SIGKILL",
-        ))
     }
 
     #[test]
@@ -793,7 +666,7 @@ mod tests {
         let script =
             create_env_script_in_dir(&dir, "echo \"$FOO\"", &[("FOO", secret)], true).unwrap();
         let invocation = script_invocation(script.path().unwrap()).unwrap();
-        let command = build_exec_command(&invocation, true);
+        let command = build_shell_command(&invocation, true);
         let argv = std::iter::once(command.get_program().to_string_lossy().to_string())
             .chain(
                 command
@@ -999,21 +872,21 @@ mod tests {
     }
 
     #[test]
-    fn truncate_preview_short_string() {
+    fn truncate_command_preview_short_string() {
         let s = "echo hello";
-        assert_eq!(truncate_preview(s), "echo hello");
+        assert_eq!(truncate_command_preview(s), "echo hello");
     }
 
     #[test]
-    fn truncate_preview_exact_limit() {
+    fn truncate_command_preview_exact_limit() {
         let s = "x".repeat(COMMAND_PREVIEW_MAX_LEN);
-        assert_eq!(truncate_preview(&s), s);
+        assert_eq!(truncate_command_preview(&s), s);
     }
 
     #[test]
-    fn truncate_preview_over_limit() {
+    fn truncate_command_preview_over_limit() {
         let s = "y".repeat(COMMAND_PREVIEW_MAX_LEN + 50);
-        let result = truncate_preview(&s);
+        let result = truncate_command_preview(&s);
         // Single-byte ASCII: truncates to exactly COMMAND_PREVIEW_MAX_LEN + "..."
         assert_eq!(
             result,
@@ -1022,30 +895,30 @@ mod tests {
     }
 
     #[test]
-    fn truncate_preview_multibyte_utf8() {
+    fn truncate_command_preview_multibyte_utf8() {
         // Each '🔥' is 4 bytes. Fill to just over the limit.
         let emoji = "🔥".repeat(COMMAND_PREVIEW_MAX_LEN / 4 + 5);
-        let result = truncate_preview(&emoji);
+        let result = truncate_command_preview(&emoji);
         assert!(result.ends_with("..."));
         // Must not panic from slicing in the middle of a UTF-8 sequence
         assert!(result.is_char_boundary(result.len() - 3));
 
         let boundary = format!("{}🔥tail", "a".repeat(COMMAND_PREVIEW_MAX_LEN - 4));
         assert_eq!(
-            truncate_preview(&boundary),
+            truncate_command_preview(&boundary),
             format!("{}🔥...", "a".repeat(COMMAND_PREVIEW_MAX_LEN - 4))
         );
 
         let crossing = format!("{}🔥tail", "a".repeat(COMMAND_PREVIEW_MAX_LEN - 1));
         assert_eq!(
-            truncate_preview(&crossing),
+            truncate_command_preview(&crossing),
             format!("{}...", "a".repeat(COMMAND_PREVIEW_MAX_LEN - 1))
         );
     }
 
     #[test]
-    fn build_exec_command_normal() {
-        let cmd = build_exec_command("echo hello", false);
+    fn build_shell_command_normal() {
+        let cmd = build_shell_command("echo hello", false);
         let prog = cmd.get_program().to_string_lossy().to_string();
         let args: Vec<String> = cmd.get_args().map(|a| a.to_string_lossy().into()).collect();
         // In debug builds: sh -c "echo hello"
@@ -1058,8 +931,8 @@ mod tests {
     }
 
     #[test]
-    fn build_exec_command_sudo() {
-        let cmd = build_exec_command("reboot", true);
+    fn build_shell_command_sudo() {
+        let cmd = build_shell_command("reboot", true);
         let prog = cmd.get_program().to_string_lossy().to_string();
         let args: Vec<String> = cmd.get_args().map(|a| a.to_string_lossy().into()).collect();
         // In debug builds: sudo sh -c "reboot"
@@ -1069,72 +942,5 @@ mod tests {
                 || (prog == "sh" && args == ["-c", "reboot"]),
             "unexpected sudo command: {prog} {args:?}"
         );
-    }
-
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn spawn_in_own_process_group_timeout_kills_background_child() {
-        let (dir, _guard) = temp_dir("pg");
-        let ready = dir.join("ready");
-        let background_pid = dir.join("background-pid");
-        let ready_arg = shell_escape_value(ready.to_str().unwrap());
-        let background_pid_arg = shell_escape_value(background_pid.to_str().unwrap());
-        let script = format!(
-            "trap '' HUP; tail -f /dev/null & echo $! > {background_pid_arg}; \
-             touch {ready_arg}; wait"
-        );
-
-        let mut command = build_exec_command(&script, false);
-        command.stdout(Stdio::null()).stderr(Stdio::null());
-        let mut child = Some(spawn_in_own_process_group(&mut command).unwrap());
-        if !wait_for_path(&ready, Duration::from_secs(2)) {
-            kill_spawned_child(&mut child);
-            panic!("background child should be started before timeout kill is tested");
-        }
-        let background_pid_text = match std::fs::read_to_string(&background_pid) {
-            Ok(pid) => pid,
-            Err(e) => {
-                kill_spawned_child(&mut child);
-                panic!("failed to read background pid: {e}");
-            }
-        };
-        let background_pid: libc::pid_t = match background_pid_text.trim().parse() {
-            Ok(pid) => pid,
-            Err(e) => {
-                kill_spawned_child(&mut child);
-                panic!("failed to parse background pid {background_pid_text:?}: {e}");
-            }
-        };
-        let background_pidfd = match open_pidfd(background_pid) {
-            Ok(pidfd) => pidfd,
-            Err(e) => {
-                kill_spawned_child(&mut child);
-                panic!("failed to open pidfd for pid {background_pid}: {e}");
-            }
-        };
-
-        let outcome = wait_with_kill_timeout(child.take().unwrap(), 100);
-        if !matches!(outcome, WaitOutcome::TimedOut) {
-            kill_pidfd_and_wait(&background_pidfd)
-                .unwrap_or_else(|e| panic!("failed to clean up background pidfd: {e}"));
-            panic!("expected timeout kill to return WaitOutcome::TimedOut");
-        }
-
-        match wait_for_pidfd_exit(&background_pidfd, Duration::from_secs(2)) {
-            Ok(true) => {}
-            Ok(false) => {
-                kill_pidfd_and_wait(&background_pidfd)
-                    .unwrap_or_else(|e| panic!("failed to clean up background pidfd: {e}"));
-                panic!(
-                    "timeout kill should terminate background pid {background_pid} in the process group"
-                );
-            }
-            Err(e) => {
-                let cleanup = kill_pidfd_and_wait(&background_pidfd);
-                panic!(
-                    "failed to wait for background pid {background_pid} exit: {e}; cleanup={cleanup:?}"
-                );
-            }
-        }
     }
 }
